Add support for `clear` command
On `clear`, the process prints ansi control codes to move the cursor to
the top left of the terminal, and remove all content after the current
cursor position

Implement enough of an ansi parser to handle these control codes

Push each chunk of data read from the shell through the ansi parser.
If we see normal data, treat it like before. If we see some form of
terminal control request, perform that instead