Fix crash on unallocated cursor write
It is valid for a program to move the cursor anywhere within the
visible viewport and write a character. This was not handled correctly
in any case except for if there was already text at the desired position

Add a TerminalBuffer struct, which acts as a grid of characters for the
cursor. Note that this is _not_ an actual grid of characters. This is
stored as the original inserted data, newlines at all, and wrapped
according to a given width to convert cursor positions into buffer
positions. We extend lines and add newlines as necessary to facilitate
the requested writes

This scheme is important to handle terminal resize in the future. When
testing other terminals, it looks like resizes result in re-wrapping.
Data that previously wrapped becomes unwrapped, and only preserves
actual newline characters. When we write past the original buffer in Y,
these newlines are preserved. Storing newlines and re-calculating is the
simplest way to get this behavior

Writes where there is existing data result in overwriting previous data

It is likely that calculating the line ranges on each insertion is
expensive, we can recalculate only on resize in the future if necessary