Bundle terminfo definition
Terminal behavior varies significantly depending on which capabilities
we claim we support. For example, removing the "clear" definition
results in "clear" not clearing the screen

Previously we were inheriting expectations on behavior of whatever
terminal launched us. This is not valid, and will result in inconsistent
behavior when launched from different sources

Copy the "dumb" terminfo definition (dumped with infocmp) and embed it
in the binary. Extract it to a temporary location for use while the
terminal is running, add back the `clear` definition and 8 color support