Set terminal width/height and send to child process
Applications expect to be able to predict the size of the terminal
emulator. This allows applications to pad things to that they fit a
given width, update the entire screen for an interactive application,
etc.

Hardcode the width and height (for now). Ensure that the GUI is
presenting a width/height that match what the application expects