Wrap content in UI at content width
Previously egui was attempting to intelligently wrap for human eyes.
E.g. on a long line with a short and long word, it would insert a
newline at the space to avoid interrupting the long word. This does
not match the way we handle wrapping internally in our emulator, which
results in a mismatch of the expected cursor position and where a human
would place it in the gui