Fix cursor rendering
UI code was attempting to parse the data in the terminal buffer, figure
out where each line was, and move the cursor from the bottom of the
label up.

This is stupid, we already handle all of this internally in the terminal
emulator. This results in incorrect cursor position when there are long
lines with wrapping

Calculate the cursor position relative to the top of the label, and just
use the offsets provided by the terminal emulator struct