Add handling of backspace output
On long lines, bash will insert backspaces to move the cursor to the
left. Handle this appropriately, and update the terminfo definition to
explicitly say we support it

Note that we do not have to handle wrapping unless terminfo claims we
do. Since we do not claim to have that behavior, we do not have to worry
about it