Add a logger
There's things I'm adding printlns for very often, and it's easy to lose
track of them. Add a logger so we can leave them in forever

Opting to avoid log crate for now. I'm assuming for now that I can do
something simpler with less effort. The log crate is so general, and
we're capable of writing one ourselves. We also have a rough case of NIH
that is hard to please. If we decide later that it's not worth it, it's
easy to upgrade in the future

Add a debug print for incoming terminal data