Implement ich ecma command (insert characters)
Readline (bash) uses the termcap "IC" to insert between other
characters. This maps to terminfo "ich", which maps to the ICH section
of ECMA-48 (8.3.64)

Implement this command. From reading spec and testing in konsole, it
seems that it should
* Insert spaces at the cursor position, pushing all text forwards
* If any text hits the end of the line, it should be deleted. I.e.
  future lines should not be effected
* It should not distinguish between line ends due to newline, or line
  ends due to wrapping

Note that end of screen behavior is undefined in ECMA-48, so those
requirements are deduced from trying another terminal emulator