Implement erase line forwards
Readline (and therefore bash) uses the erase in line command when
updating the current input line. Implement termcap "ce", which is
terminfo "el", which is ecma 8.3.41.

When this command is executed, we should remove from the current cursor
position to the end of the line. Note that we do not handle other
deletion modes (up to cursor, or full line), because it was not
necessary for bash's usage. This is all we care about at this point