Add E3 command to terminfo
Readline supports a keybind to clear the entire scrollback
(clear-display). This uses readline to fetch the E3 termcap and puts it
on the terminal.

This is a user capability (man user_caps), so tell the tic compiler to
include user capabilities when compiling and add the capability. This is
actually how our regular clear works right now and our code already
treats clear modes 2 and 3 (xterm extension to clear scrollback) the
same. No actual implementation is needed