Fix color tracking when inserting text in middle of buffer
In the previous implementation, we did not account for the fact that
there was a difference between inserting data and overwriting what was
there, and inserting new data. This resulted in color information
getting attached to the wrong characters if we shifted the buffer
position of a character

Track insertions separately from written data. This allows us to adjust
the format ranges as we need to keep color tracking data in tact

An example of the previous error

```bash
echo -e "\e[31masdf\e[m\e[3G\e[4@1234"
```

This should show "as" in red, followed by "1234" in default, and "df" in
red again. With the previous implementation this would not work as
expected