Fix terminal_buffer commands not relative to visible area
In 2d640f9840a446b15d8620e2d56949178fc47ab9 we added the concept of
"visible line ranges" to distinguish between what is currently on
screen and what's in the scrollback. Unfortunately in this commit we
forgot to make this distinction in all relevant areas.

Correct this now and ensure that all functions in TerminalBuffer use the
visible area for cursor position adjustment

With the changes to clear_fowrwards, append_newline_at_line_end becomes
unneeded, as every line has a newline appended at all times by default

Bonus fixes:
* In calc_line_ranges, the range could end up being too big. This was a
  result of double counting the newline
* Change last_line_idx -> num_lines, a more correct name as the
  last_line_idx is len() - 1
* Occasionally the format tracker ends up tracking the end of the buffer
  (a newline), which gets stripped by the gui to avoid incorrect padding
  in the UI. This results in an out of bounds access caused by format
  data. Handle this case more gracefully