Implement insert lines
Implement ecma-48 8.3.67, the insert lines command.

Less complains about missing features, this is one of the features it
requires. The feature is supposed to take the line that the cursor is
on, and create space for N lines below to be inserted. Lines below are
scrolled down, and if they hit the edge of the screen they are removed

* Add ansi parser for 'L'
* Add insert_lines command to terminal buffer
* Add terminfo definitions

Format tracker integration was pretty straight forward. Re-use the
existing APIs and just track deleted and inserted buffer positions like
we did for inserting spaces