Fix crash when format data is on the edge of the buffer
In some cases the format data is right on the edge of the buffer. E.g.
if the buffer length is 10, there may be format data at [0..10] and
[10..usize::max]. If the last character is a newline, the gui code will
strip it to avoid extra padding, and now the whole segment is out of
range.

1. Remove the warning log in this case, it is reasonably expected now
2. Heal the end of the format range as well, as the end of the range may
   perfectly line up with the end of the buffer, which has been deleted

Repro was to try the following at various font sizes and window sizes.
```
echo -ne "\e[31mtest\e[m\e[5L\r"
```