Abstract terminal IO away from terminal emulator impl
In a future commit, we are going to add the ability to replay recordings
that we took of some program. In this scenario we do not want to
interact with a file descriptor of a pty, as there should be no pty
running. In the case of a recording we will want to ask the replay
system if there is data to consume.

Create a new trait, TermIO. This should take anything that interacts
with linux, and hide it away