Move gui to module folder
In a future commit, we will be splitting out most of the terminal logic
into it's own file. This will need gui to be a folder so that we can
have multiple widgets/files in the module. Do the move now with only the
required changes to keep compiling to make future diffs easier to parse